cassandra-benchmark/requirements.txt
# Cassandra TPC-C Benchmark Framework Dependencies

# Cassandra Database Driver
cassandra-driver>=3.25.0

# Configuration Management
pyyaml>=6.0

# CLI Interface
click>=8.1.0

# Console Output Formatting
tabulate>=0.9.0

# Statistical Calculations
numpy>=1.24.0

# Data Analysis (optional)
pandas>=2.0.0

# Metrics Export (optional)
prometheus-client>=0.17.0

# Type Checking Support
typing-extensions>=4.5.0
